requirements.txt
gitpython
paramiko